requirements.txt
click==8.1.3
matplotlib==3.6.2
numpy==1.23.5
omegaconf==2.3.0
python-dotenv==0.21.0
pytorch_lightning==1.8.6
setuptools==65.6.3
torch==1.13.1
